requirements.txt
PyYAML==6.0.1
jsonschema==4.23.0
pyarrow==19.0.1
click==8.1.8 